Improve test running and watching for typescript.
This adds some helper scripts and nodemon to watch the directory and run
tests in a robust way while watching for changes.

We experimented with `ts-jest` and similar testing methods that would
watch and run tests directly from their `test` or `src` directories, but
this was proving to have weird edge cases where sometimes files wouldn't
recompile.

At the end of the day, I want the tests running against the *exact* code
we're shipping, so the chosen method is a little slower (in that it
always does a full rebuild) but is robust to anomalies and edge cases.